reset: Make reset command state transitory
When a reset command is sent to the device, it should go back to the
"off" state automatically, as there is no such binary state in practice.
However, this automatic going down should not trigger another "reset"
output to the device, as it would happen by simply adding a `HIGH`
field.

Therefore, the approach used here is to use a `calcout` record only to
conditionally send the output. It directly maps the transitory state
from `RFFERst-Cmd` to the condition (which is also the output value to
the protocol, but not used at all), and then uses it to define that the
output should be sent when the reset is in the `on` state. This prevents
triggering the output when the `HIGH` timeout expires.