attenuation: Add attenuation setting support
Read-back record has been updated to use periodic scanning. `I/O Intr`
was previously being used to enable the underlying driver to pass the
upcoming data to the application even when the processing state is not
expecting it. With StreamDevice this should no longer be needed as
forward links will be used. The scan time has been chosen arbitrarily to
be a reasonable value not to overload the device nor keep the read-back
value eventually outdated for long.

Since forwarding is to a non-passive record, the `PROC` field is the
target of the forwarding field using the channel access (CA) protocol.

Moreover, when the IOC is started, attenuation read-back entry should be
updated to the value currently set in the hardware, so that it is
consistent from the start-up. Therefore, a `PINI` has been set. For the
set-point record, PINI has been kept from the original IOC.