temperature: Add device temperature monitoring support
Monitor parameters ADEL and MDEL configured to -1 have been removed from
the previous implementation to this new one. This is because ADEL and
MDEL is usually argued to be needed to identify whether an anomalous
event has occurred and the IOC did not report anything (since the value
did not change). However, this is not a sufficient argument as an alarm
will be triggered by EPICS showing that an event has occurred. For
instance, when a timeout occurs, it will be displayed in the `camonitor`
output, and will also be registered by Archiver (since it also listens
to these alarms).

The scan time has been kept from the previous implementation. It might
be investigated whether it is a good option to increase the data points
by lowering the scan period while avoiding overloading the device.