asyn: Configure asynRecord for device connection monitoring
In the current deploy setting, there are RFFE IOC clients that monitor
the device up status by querying the `CNCT` field from asynRecord. This
has been decided to be kept for the new IOC. PV naming as well as IMAX
and OMAX parameters have been inherited from the previous IOC
implementation.

However, for this to work as expected, we should set up the
`disconnectOnReadTimeout` asyn option introduced in R4-29 [1], and
patched for StreamDevice compatibility in R4-35 [2]. Without this
option, asyn will only close the socket (and therefore update the CNCT
PV value) when TCP connection is closed due to successive retransmission
timeouts. More specifically, in the default Linux configuration, TCP
will attempt to retransmit the segment for 15 times (with increasing
timeouts) before giving up as defined by the
`/proc/sys/net/ipv4/tcp_retries2` config, which means about 15 minutes
in a local network.

With disconnection on read timeout, we must ensure that autoConnect
option is set up, so that the IOC can automatically reconnect to the
device when a delayed response eventually causes a read timeout [3].
Since R2-1, this is enabled by default.

[1]: https://epics.anl.gov/tech-talk/2019/msg00290.php
[2]: https://epics-modules.github.io/master/asyn/R4-35/RELEASE_NOTES.html
[3]: https://epics.anl.gov/tech-talk/2019/msg00293.php