docker: Containerize IOC
EPICS and its modules were previously configured based on the versions
available in the Sirius NFS server. Since these are now built inside the
container, EPICS base has been updated to its later 3.15 version:
3.15.9. Each module has been also updated to the latest version.

Note that `calc` has been removed. StreamDevice was built with `calc` in
the NFS system, and thus required `calc` to be loaded in the IOC. Since
we do not actually use it in RFFE IOC, it has been now removed both from
the StreamDevice and RFFE IOC dependencies.

Note that the directories in the RELEASE file does not show the version
number of each module. This has been used to make it easier to specify
the dependencies in other RELEASE files. The versions are thus
documented in `docker/Dockerfile` only.

Build artifacts must not be copied by Docker engine inside the image
when building it, since it may break the installation scripts.
Therefore, a `.dockerignore` has been configured similarly to the
`.gitignore` by adapting the Git patterns to the ones supported by
Docker. Moreover, Git repository is available by other means with
unchanged history, thus it has been removed from the resulting image
along with its configuration files (e.g. .gitignore).

All interactive options from `apt` have been disabled so that the script
does not fail asking for confirmation. This includes the `apt update`,
even tough currently it seems not be using this interactive prompts.