init: Use procServ to start up IOC
The IOC is now initialized with procServ with the correct handling of PV
prefixes to be used in deploy. Instead of declaring area and device
prefixes by hand, we specify a mapping file with environment variables
for each crate and BPM slot, as it was done before along with the init
system. The needed prefixes are then exported by the `run.sh` script
with the expected name for the IOC. This is done through BASH indirect
expansion, which basically gets the value of the variable specified by
another variable. In this case, `{area,device}_prefix_var` specifies
which variable to use from the `slot-mapping.env` file.

Instead of using an EPICS_PV_CRATE_PREFIX to define the crate, simply a
CRATE_NUMBER is now required to be specified. This makes it easier to
specify the crate as all env variables are prefixed with `CRATE_`.

Unix socket has been configured for procServ to avoid conflicts of ports
in the host, since the `--network host` option is configured. Since the
socket is needed during runtime, the directory is created in the
`run.sh` instead of the Dockerfile. The path used by the socket file
follows the specified by the Filesystem Hierarchy Standard 3.0.