autosave: Set up automatic save and restore
This configuration follows the general settings that were used in the
previous IOC implementation. This includes the PVs that are handled by
autosave, as previously specified by the
`BPMRFFEApp/Db/BPMRffe_settings.req`. This file has been renamed to
`bpmrffe.req` in the current implementation as no other requirements
file exists for this IOC as of now, and it matches the database file
name that stores the specified records.

Moreover, the indirect loading from `auto_settings.req` to
`BPMRFFE_settings.req`, and then to `BPMRffe_settings.req`, have been
removed, as they introduce complexity while not bringing any benefit.
This means that monitor is now created directly for the `bpmrffe.req`
file.

Autosave has been configured to save its `.sav` files in a subdirectory
of `/var/opt/rffe-epics-ioc`, which is a FHS 3.0 compliant path. This
directory should be bound to the host filesystem so it can be persisted.
Each IOC will store its files named after their prefix, so that there is
no conflict for a shared filesystem binding.