init: Configure RFFE IP address from given base address
In the current deploy setting, RFFEs have sequential IP addresses in
each crate. Therefore, for making it easier to deploy, each container
computes its RFFE IP address based on the `BPM_NUMBER` and
`RFFE_BASE_IP_ADDRESS` environment variables.

A simplification has been made in the current implementation so that
only the last dotted-decimal part of the address is offset based on the
BPM_NUMBER. In practice, this means the last valid base address ends
with 255 - 24 (max slot number) = 231 if we consider the last address
(ending with 255) is not a broadcast address.