deploy: Add docker-compose to launch IOCs
This configuration follows the current deploy setting at Sirius. Some
key points here are:

Compose file version 3.2 is used to allow us to bind mount the autosave
directory [1]. It works with the version of docker-compose we have.

Several different services are used to make it possible to have a
different container for each BPM slot (and RFFE). Thus, a base
configuration has been set and is inherited for each service. However,
using the base service for the environment and then appending to it
inside a service doesn't work in the version of docker-compose we have
on our crates, so we need to declare the full environment in each one.

In our current configuration, all crates use the same local network
addresses for RFFEs. This means we can include this configuration
directly in this configuration file.

Co-authored-by: Érico Nogueira <erico.rolim@lnls.br>

[1] https://github.com/docker/compose/issues/4763