attenuation: Limit output value to [0, 31.5]dB interval
The current firmware version (v3.0.0) does not check for invalid
attenuation inputs. This leads to the read-back value being set to the
any given value, even though the hardware itself will truncate the MSB
outside the range of the attenuator. In order to prevent such
misinformation in the IOC PVs, the output value has been is now clipped
to [0, 31.5]dB.