docker: upgrade to epics-in-docker v0.3.0.
Since `RELEASE` is configured during the build, there is not need to
define the paths in this repository. Thus, to avoid definition
misinterpretation, all paths have been left blank. All dependend modules
still documented in the RELEASE file, as it is a standard place to look
for dependencies.

Target build stage is defined explicitly, so that it is consistent with
other IOCs and the epics-in-docker guidelines.