deploy: use unless-stopped restart policy.
This should allow Docker to restart a stopped container when udev has
not triggered `docker-compose up` in the appropriate time, which is
after Docker daemon has started. This removes the need to add the Docker
service as a dependency for starting up other modules in bpm-app.

Unless-stopped policy has been chosen instead of "always", so that we can
still have it manually stopped without removing it.