deploy: provide a systemd unit managing containers.
When attempting to run in a podman namespace through a udev-triggered
script, filesystem mounting does not work out-of-the-box. When
configured with overlayfs, a permission denied occurs when remouting
`containers/storage/overlay/` with flags 0x40000. Similar issues happen
with vfs, and probably other filesystems.

Launching the container under systemd does not provoke such not
understood issues. Let's provide a systemd-unit to be started by udev
when the IOC must be run. This is similar to the AFC EPICS IOC deploy,
also making services management uniform.

During stop, a timeout of zero is provided to use SIGKILL right away,
since the SIGTERM signal does not reach the procServ process and its
childs for a yet unknown reason.