deploy: remove containerized deployment.
We are no longer going to use containers for this application, because
we are interested in moving away from Docker, and launching Podman from
udev rules, as our design requires, no longer works on modern distros
[1]. Launching a service to monitor podman would work, but requires
complicated integration (the native systemd-podman integration isn't
available in Debian 12), so we decided to move to the same setup as
afc-epics-ioc, with static binaries.

The container creation scheme is also removed.

[1] https://www.marcusfolkesson.se/blog/systemd-udev/